Finished booking page

+++ src/main/java/com/example/repository/VaccineAppointmentRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface VaccineAppointmentRepository extends JpaRepository<VaccineAppointment, Long> {}
+public interface VaccineAppointmentRepository extends JpaRepository<VaccineAppointment, Integer> {}
